requirements.txt
streamlit
pandas
pymysql
plotly
google-generativeai
groq
python-dotenv
scikit-learn
mlxtend
html2text
openpyxl
xgboost
requests
beautifulsoup4
numpy
